app/requirements4.txt
django==2.2.6
psycopg2==2.8.3
psycopg2-binary==2.8.3
Pillow==6.1.0
django-crispy-forms==1.7.2
stripe==2.37.2
ib_insync==0.9.56
SQLAlchemy==1.3.5
celery==3.1.24
redis==2.10.3
djangorestframework==3.10.3
django-rest-auth==0.9.5
django-allauth==0.40.0
httpie==1.0.3

